+++ CHANGELOG.md
@@ -33,3 +33,8 @@
 
 ## 0.2.3
 - Create data grid buttons control to centralized code
+
+## 0.3.0
+- Fix problem with Bible Fact Pak™ import where existing items will be re-imported a second time
+- Refactor IQuestionImporter interface so default constructors can now be used
+- Refactor Bible Fact-Pak™ import user interface into a generic 'direct text import' user control